requirements.txt
Flask==2.3.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai
python-dotenv==1.0.0
numpy==1.24.3